+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-3.0.0 - TBD
+3.0.0 - June 2023
 -----------------
 * Drop support for PHP 7.4
 * Update minimum supported Yii version to the minimum that supports PHP 8.0